# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Business Manager API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Item {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  category: string;
  unitPrice: number;
  costPrice: number;
  unitOfMeasure: string;
  createdAt: string;
  updatedAt?: string;
}

export interface ItemInput {
  name: string;
  description?: string;
  category: string;
  unitPrice: number;
  costPrice: number;
  unitOfMeasure: string;
}

export interface ItemUpdate {
  name?: string;
  description?: string;
  category?: string;
  unitPrice?: number;
  costPrice?: number;
  unitOfMeasure?: string;
}

export interface InventoryLevel {
  itemId: number;
  quantityOnHand: number;
  reorderPoint: number;
  updatedAt: string;
  item?: Item;
}

export interface InventoryLevelUpdate {
  quantityOnHand?: number;
  reorderPoint?: number;
}

export interface StockAdjustment {
  id: number;
  itemId: number;
  quantity: number;
  reason: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface StockAdjustmentInput {
  quantity: number;
  reason: string;
  notes?: string;
}

export interface Customer {
  id: number;
  name: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  country?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface CustomerInput {
  name: string;
  email?: string;
  phone?: string;
  address?: string;
  city?: string;
  country?: string;
  notes?: string;
}

export interface CustomerUpdate {
  name?: string;
  email?: string;
  phone?: string;
  address?: string;
  city?: string;
  country?: string;
  notes?: string;
}

export interface Supplier {
  id: number;
  name: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  country?: string | null;
  /** @nullable */
  contactPerson?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface SupplierInput {
  name: string;
  email?: string;
  phone?: string;
  address?: string;
  city?: string;
  country?: string;
  contactPerson?: string;
  notes?: string;
}

export interface SupplierUpdate {
  name?: string;
  email?: string;
  phone?: string;
  address?: string;
  city?: string;
  country?: string;
  contactPerson?: string;
  notes?: string;
}

export interface InvoiceLineItem {
  id: number;
  invoiceId: number;
  /** @nullable */
  itemId: number | null;
  description: string;
  quantity: number;
  unitPrice: number;
  lineTotal: number;
  item?: Item;
}

export type InvoiceStatus = typeof InvoiceStatus[keyof typeof InvoiceStatus];


export const InvoiceStatus = {
  Draft: 'Draft',
  Sent: 'Sent',
  Paid: 'Paid',
  Overdue: 'Overdue',
} as const;

export interface Invoice {
  id: number;
  invoiceNumber: string;
  customerId: number;
  status: InvoiceStatus;
  /** @nullable */
  issueDate?: string | null;
  /** @nullable */
  dueDate?: string | null;
  subtotal: number;
  taxRate: number;
  taxAmount: number;
  total: number;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  customer?: Customer;
  lineItems?: InvoiceLineItem[];
}

export interface InvoiceLineItemInput {
  itemId?: number;
  description: string;
  quantity: number;
  unitPrice: number;
}

export interface InvoiceInput {
  customerId: number;
  issueDate?: string;
  dueDate?: string;
  taxRate?: number;
  notes?: string;
  lineItems: InvoiceLineItemInput[];
}

export interface InvoiceUpdate {
  customerId?: number;
  issueDate?: string;
  dueDate?: string;
  taxRate?: number;
  notes?: string;
  lineItems?: InvoiceLineItemInput[];
}

export type InvoiceStatusUpdateStatus = typeof InvoiceStatusUpdateStatus[keyof typeof InvoiceStatusUpdateStatus];


export const InvoiceStatusUpdateStatus = {
  Draft: 'Draft',
  Sent: 'Sent',
  Paid: 'Paid',
  Overdue: 'Overdue',
} as const;

export interface InvoiceStatusUpdate {
  status: InvoiceStatusUpdateStatus;
}

export interface PurchaseOrderLineItem {
  id: number;
  purchaseOrderId: number;
  /** @nullable */
  itemId: number | null;
  description: string;
  quantity: number;
  unitCost: number;
  lineTotal: number;
  item?: Item;
}

export type PurchaseOrderStatus = typeof PurchaseOrderStatus[keyof typeof PurchaseOrderStatus];


export const PurchaseOrderStatus = {
  Draft: 'Draft',
  Sent: 'Sent',
  Received: 'Received',
  Cancelled: 'Cancelled',
} as const;

export interface PurchaseOrder {
  id: number;
  poNumber: string;
  supplierId: number;
  status: PurchaseOrderStatus;
  /** @nullable */
  orderDate?: string | null;
  /** @nullable */
  expectedDate?: string | null;
  subtotal: number;
  total: number;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  supplier?: Supplier;
  lineItems?: PurchaseOrderLineItem[];
}

export interface PurchaseOrderLineItemInput {
  itemId?: number;
  description: string;
  quantity: number;
  unitCost: number;
}

export interface PurchaseOrderInput {
  supplierId: number;
  orderDate?: string;
  expectedDate?: string;
  notes?: string;
  lineItems: PurchaseOrderLineItemInput[];
}

export interface PurchaseOrderUpdate {
  supplierId?: number;
  orderDate?: string;
  expectedDate?: string;
  notes?: string;
  lineItems?: PurchaseOrderLineItemInput[];
}

export type PurchaseOrderStatusUpdateStatus = typeof PurchaseOrderStatusUpdateStatus[keyof typeof PurchaseOrderStatusUpdateStatus];


export const PurchaseOrderStatusUpdateStatus = {
  Draft: 'Draft',
  Sent: 'Sent',
  Received: 'Received',
  Cancelled: 'Cancelled',
} as const;

export interface PurchaseOrderStatusUpdate {
  status: PurchaseOrderStatusUpdateStatus;
}

export interface DashboardSummary {
  totalInventoryValue: number;
  outstandingInvoicesTotal: number;
  outstandingInvoicesCount: number;
  lowStockCount: number;
  totalCustomers: number;
  totalSuppliers: number;
  totalItems: number;
  paidInvoicesThisMonth?: number;
  pendingPurchaseOrders?: number;
}

export type ActivityItemType = typeof ActivityItemType[keyof typeof ActivityItemType];


export const ActivityItemType = {
  invoice: 'invoice',
  purchase_order: 'purchase_order',
  stock_adjustment: 'stock_adjustment',
  customer: 'customer',
  supplier: 'supplier',
} as const;

export interface ActivityItem {
  id: string;
  type: ActivityItemType;
  description: string;
  /** @nullable */
  amount?: number | null;
  /** @nullable */
  status?: string | null;
  createdAt: string;
}

export interface InvoiceStats {
  draft: number;
  sent: number;
  paid: number;
  overdue: number;
  totalDraft?: number;
  totalSent?: number;
  totalPaid?: number;
  totalOverdue?: number;
}