# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Business Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /inventory/{itemId}:
    get:
      operationId: getInventoryLevel
      tags: [inventory]
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryLevel"
    patch:
      operationId: updateInventoryLevel
      tags: [inventory]
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryLevelUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryLevel"
  /inventory/{itemId}/adjust:
    post:
      operationId: adjustStock
      tags: [inventory]
      summary: Record a stock adjustment
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StockAdjustmentInput"
      responses:
        "201":
          description: Adjustment recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockAdjustment"
  /inventory/{itemId}/adjustments:
    get:
      operationId: listStockAdjustments
      tags: [inventory]
      summary: Adjustment history for an item
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StockAdjustment"
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    patch:
      operationId: updateCustomer
      tags: [customers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── SUPPLIERS ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Item:
      type: object
      required: [id, name, category, unitPrice, costPrice, unitOfMeasure, createdAt]
      properties:
        id:            { type: integer }
        name:          { type: string }
        description:   { type: ["string", "null"] }
        category:      { type: string }
        unitPrice:     { type: number }
        costPrice:     { type: number }
        unitOfMeasure: { type: string }
        createdAt:     { type: string }
        updatedAt:     { type: string }
...
    InventoryLevel:
      type: object
      required: [itemId, quantityOnHand, reorderPoint, updatedAt]
      properties:
        itemId:          { type: integer }
        quantityOnHand:  { type: number }
        reorderPoint:    { type: number }
        updatedAt:       { type: string }
        item:
          $ref: "#/components/schemas/Item"
    InventoryLevelUpdate:
      type: object
      properties:
        quantityOnHand: { type: number }
        reorderPoint:   { type: number }
    StockAdjustment:
      type: object
      required: [id, itemId, quantity, reason, createdAt]
      properties:
        id:        { type: integer }
        itemId:    { type: integer }
        quantity:  { type: number }
        reason:    { type: string }
        notes:     { type: ["string", "null"] }
        createdAt: { type: string }
    StockAdjustmentInput:
      type: object
      required: [quantity, reason]
      properties:
        quantity: { type: number }
        reason:   { type: string }
        notes:    { type: string }

    # ── CUSTOMER ─────────────────────────────────────────────────────────────
    Customer:
      type: object
      required: [id, name, createdAt]
      properties:
        id:        { type: integer }
        name:      { type: string }
        email:     { type: ["string", "null"] }
        phone:     { type: ["string", "null"] }
        address:   { type: ["string", "null"] }
        city:      { type: ["string", "null"] }
        country:   { type: ["string", "null"] }
        notes:     { type: ["string", "null"] }
        createdAt: { type: string }
    CustomerInput:
      type: object
      required: [name]
      properties:
        name:    { type: string }
        email:   { type: string }
        phone:   { type: string }
        address: { type: string }
        city:    { type: string }
        country: { type: string }
        notes:   { type: string }
    CustomerUpdate:
      type: object
      properties:
        name:    { type: string }
        email:   { type: string }
        phone:   { type: string }
        address: { type: string }
        city:    { type: string }
        country: { type: string }
        notes:   { type: string }

    # ── SUPPLIER ─────────────────────────────────────────────────────────────
...