# lib/api-zod/src/generated/types/item.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Business Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface Item {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  category: string;
  unitPrice: number;
  costPrice: number;
  unitOfMeasure: string;
  createdAt: string;
  updatedAt?: string;
}